+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This list is not currently intended to be all-encompassing - it will document major and breaking API changes with their rationale when appropriate:
 
 ### v3.115.0 (uncut)
+- Chaos now do not blat `x-uri-template` when used with a `RoutingHttpHandler`
 - Simplified usage of `Once` chaos trigger.
 - (Slight break) Consistentified (!) construction of Chaos Behaviours, Stages and Triggers. Replaced 
 singletons with function calls. Eg. `Always -> Always()`